+++ tomcat/src/main/java/nextstep/jwp/servlets/RootServlet.java
@@ -1,10 +1,8 @@
-package org.apache.catalina.servlets;
+package nextstep.jwp.servlets;
 
-import static org.apache.coyote.http11.PagePathMapper.INDEX_PAGE;
 import static org.apache.coyote.http11.message.HttpHeaders.CONTENT_LENGTH;
 import static org.apache.coyote.http11.message.HttpHeaders.CONTENT_TYPE;
 
-import java.io.IOException;
 import org.apache.coyote.http11.ContentType;
 import org.apache.coyote.http11.message.HttpMethod;
 import org.apache.coyote.http11.message.HttpStatus;
@@ -12,10 +10,10 @@ import org.apache.coyote.http11.message.request.HttpRequest;
 import org.apache.coyote.http11.message.response.HttpResponse;
 import org.apache.coyote.http11.message.response.ResponseBody;
 
-public class IndexServlet extends Servlet {
+public class RootServlet extends Servlet {
 
     @Override
-    public void service(HttpRequest httpRequest, HttpResponse httpResponse) throws IOException {
+    public void service(HttpRequest httpRequest, HttpResponse httpResponse) {
         HttpMethod httpMethod = httpRequest.getMethod();
         if (httpMethod.isEqualTo(HttpMethod.GET)) {
             doGet(httpRequest, httpResponse);
@@ -23,14 +21,13 @@ public class IndexServlet extends Servlet {
     }
 
     @Override
-    protected void doGet(HttpRequest httpRequest, HttpResponse httpResponse) throws IOException {
-        String absolutePath = INDEX_PAGE.path();
-        String content = findResourceWithPath(absolutePath);
+    protected void doGet(HttpRequest httpRequest, HttpResponse httpResponse) {
+        String content = "Hello world!";
         ResponseBody responseBody = new ResponseBody(content);
 
         httpResponse.setHttpVersion(httpRequest.getHttpVersion())
                 .setHttpStatus(HttpStatus.OK)
-                .addHeader(CONTENT_TYPE, ContentType.parse(absolutePath))
+                .addHeader(CONTENT_TYPE, ContentType.parse(content))
                 .addHeader(CONTENT_LENGTH, String.valueOf(content.getBytes().length))
                 .setResponseBody(responseBody);
     }
