+++ tomcat/src/main/java/nextstep/jwp/servlets/StaticResourceServlet.java
@@ -1,4 +1,4 @@
-package org.apache.catalina.servlets;
+package nextstep.jwp.servlets;
 
 import static org.apache.coyote.http11.message.HttpHeaders.CONTENT_LENGTH;
 import static org.apache.coyote.http11.message.HttpHeaders.CONTENT_TYPE;
