+++ tomcat/src/main/java/org/apache/catalina/container/Context.java
@@ -1,9 +1,9 @@
-package org.apache.catalina.servlets;
+package org.apache.catalina.container;
 
 import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
-import org.apache.catalina.servlets.*;
+import nextstep.jwp.servlets.*;
 import org.apache.coyote.http11.message.request.HttpRequest;
 import org.apache.coyote.http11.message.request.RequestURI;
 import org.apache.coyote.http11.message.response.HttpResponse;
